requirements.txt
fastapi
uvicorn
openai
azure-search-documents
azure-identity
azure-storage-blob
python-dotenv
pypdf